dalle_mini/requirements.txt
absl-py==0.13.0
appnope==0.1.2
argcomplete==1.12.1
argon2-cffi==21.1.0
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
blobconverter==0.0.10
boto3==1.17.84
botocore==1.20.84
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
chex==0.0.8
click==8.0.1
configparser==5.0.2
cycler==0.10.0
datasets==1.11.0
debugpy==1.4.1
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
dm-tree==0.1.6
docker-pycreds==0.4.0
entrypoints==0.3
filelock==3.0.12
flatbuffers==2.0
flax==0.3.4
fsspec==2021.8.1
ftfy==6.0.3
gitdb==4.0.7
GitPython==3.1.18
huggingface-hub==0.0.16
idna==2.10
ipykernel==6.3.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.4
jax==0.2.20
jaxlib==0.1.71
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.2
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.1
kiwisolver==1.3.1
libtpu-nightly==0.1.dev20210809
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.3
mistune==0.8.4
msgpack==1.0.2
multiprocess==0.70.12.2
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.4.3
numpy==1.19.5
open3d==0.10.0.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
optax==0.0.9
packaging==21.0
pandas==1.3.2
pandocfilters==1.4.3
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
pyarrow==5.0.0
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytube==10.8.1
pytz==2021.1
PyYAML==5.4
pyzmq==22.2.1
qtconsole==5.1.1
QtPy==1.11.0
regex==2021.8.28
requests==2.24.0
s3transfer==0.4.2
sacremoses==0.0.45
scipy==1.7.1
Send2Trash==1.8.0
sentry-sdk==1.3.1
shortuuid==1.0.1
six==1.16.0
smmap==4.0.0
subprocess32==3.5.4
terminado==0.12.1
testpath==0.5.0
tokenizers==0.10.3
toolz==0.11.1
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
transformers==4.10.0
typing-extensions==3.10.0.2
urllib3==1.25.11
vqgan-jax @ git+https://github.com/patil-suraj/vqgan-jax.git@610d842dd33c739325a944102ed33acc07692dd5
wandb==0.12.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xxhash==2.0.2
requests
-f https://storage.googleapis.com/jax-releases/libtpu_releases.html
jax[tpu]>=0.2.16
transformers
datasets
flax
jupyter
wandb
nltk
optax
git+https://github.com/patil-suraj/vqgan-jax.git@610d842dd33c739325a944102ed33acc07692dd5

# Inference
ftfy

#MP4MP3
moviepy
mutagen
